pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Automated Zettelkasten research agent with Obsidian + Telegram"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "tavily-python>=0.5.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.68.0",
]

[project.scripts]
research-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "interfaces", "llm", "processing", "search", "storage"]
